Move generate to MazeAlgo trait
